+++ NEWS.md
@@ -44,6 +44,7 @@ Makie.save("surface_plot.png", fig)
 
 - The number of dimensions of a `OutputVar` can be accessed with `ndims`.
 - You can drop dimensions of size 1 with `dropdims`.
+- Add `arecompatible` for `FlatVar` and `Metadata`.
 
 ## Bug fixes